checkpoint from test stuff
