+++ CMakeLists.txt
@@ -1,28 +1,39 @@
 cmake_minimum_required(VERSION 3.16)
-################################################################################
-##              USER CONFIGURED 
-################################################################################
-# - when cross compiling is off, we build to run on our host system
-# 
-# - when cross compiling is on, we build for target device
-################################################################################
-
-
-# https://sourceforge.net/projects/mingw-w64/
 
 if(NOT DEFINED CMAKE_CROSSCOMPILING)
     message(FATAL_ERROR "CMAKE_CROSSCOMPILING IS NOT DEFINED. Please define it as \"ON\" or \"OFF\"")
     return()
 endif()
 
+if(NOT CMAKE_BUILD_TYPE)
+    set(CMAKE_BUILD_TYPE Debug)
+endif()
 
-# wget -qO- http://software-dl.ti.com/msp430/msp430_public_sw/mcu/msp430/MSPGCC/9_2_0_0/export/msp430-gcc-9.2.0.50_linux32.tar.bz2 | tar -xj
+list(APPEND SUPPORTED_BUILD_TYPES "Debug")
+list(APPEND SUPPORTED_BUILD_TYPES "Release")
+list(APPEND SUPPORTED_BUILD_TYPES "RelWithDebInfo")
+list(APPEND SUPPORTED_BUILD_TYPES "MinSizeRel")
+if(NOT ${CMAKE_BUILD_TYPE} IN_LIST SUPPORTED_BUILD_TYPES)
+    message(FATAL_ERROR "${CMAKE_BUILD_TYPE} is not a supported build type")
+    message("Supported build types:")
+    foreach(btype ${SUPPORTED_BUILD_TYPES})
+        message("- \"${btype}\"")    
+    endforeach(btype ${SUPPORTED_BUILD_TYPES})
+endif(NOT ${CMAKE_BUILD_TYPE} IN_LIST SUPPORTED_BUILD_TYPES)
 
 set(MCU_MPN "msp430f5529") #set this as a lowercase string of your part number
 set(MCU_FREQ "8000000")    #set this as the HSE input freq in Hz
 
 ################################################################################
-project(ADCS_firmware LANGUAGES C ASM)
+#                       PROJECT START                                          #
+################################################################################
+project(
+    ADCS_application 
+    VERSION 0.1
+    DESCRIPTION "ADCS Firmware Project for DSS LORIS Satellite"
+    LANGUAGES C ASM
+)
+
 message("") 
 message("PROJECT : ${CMAKE_PROJECT_NAME}")
 message("TARGET : ${CMAKE_SYSTEM_NAME}")
@@ -30,32 +41,13 @@ message("PROCESSOR : ${CMAKE_SYSTEM_PROCESSOR}")
 message("") 
 
 set(CMAKE_EXPORT_COMPILE_COMMANDS ON)
-if(NOT CMAKE_BUILD_TYPE)
-    set(CMAKE_BUILD_TYPE Debug)
-endif()
 
-if(CMAKE_BUILD_TYPE STREQUAL "Debug")
-    add_compile_options("-Wall")
-    add_compile_options("-Wshadow")
-    add_compile_options("-g")
-    add_compile_options("-ggdb")
-endif()
-
-set(CMAKE_RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR})
-set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR})
-set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR})
-
-if(CMAKE_CROSSCOMPILING STREQUAL "ON")
-    set(CMAKE_EXECUTABLE_SUFFIX ".elf")
-else()
-    if(WIN32)
-        set(CMAKE_EXECUTABLE_SUFFIX ".exe")
-    elseif(UNIX)
-        set(CMAKE_EXECUTABLE_SUFFIX ".out")
-    else()
-        message("UNSUPPORTED PLATFORM : ${CMAKE_HOST_SYSTEM_NAME}")
-    endif()
-endif()
+#if(CMAKE_BUILD_TYPE STREQUAL "Debug")
+#    add_compile_options("-Wall")
+#    add_compile_options("-Wshadow")
+#    add_compile_options("-g")
+#    add_compile_options("-ggdb")
+#endif()
 
 function(ADCS_add_executable target)
     if(CMAKE_CROSSCOMPILING STREQUAL "ON")
@@ -66,7 +58,6 @@ function(ADCS_add_executable target)
         add_executable(${target} ${ARGN})
     endif(CMAKE_CROSSCOMPILING STREQUAL "ON")
 endfunction(ADCS_add_executable target)
-
 
 function(ADCS_add_library target)
     if(CMAKE_CROSSCOMPILING STREQUAL "ON")
@@ -84,7 +75,13 @@ include_directories("${CMAKE_CURRENT_SOURCE_DIR}/shared/inc")
 # FOR THE FIRST TIME, DO 
 # $git submodule update --force --remote --recursive --init
 # NOTICE THE --init on the end.
-execute_process(COMMAND "git submodule update --force --remote --recursive")
+execute_process(
+    COMMAND "git" "submodule" "update" "--force" "--remote" "--recursive"
+    OUTPUT_VARIABLE GIT_SUBMODULE_UPDATE_OUTPUT_VARIABLE
+    RESULT_VARIABLE GIT_SUBMODULE_UPDATE_RESULT_VARIABLE
+    OUTPUT_STRIP_TRAILING_WHITESPACE
+    ECHO_OUTPUT_VARIABLE
+    )
 
 add_subdirectory(JTOK)
 add_subdirectory(drivers)
